refac: service move some methods to influx wrapper
